requirements.txt
# --- WEB FRAMEWORK & SERVER ---
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9

# --- DATA PROCESSING ---
pandas==2.2.1
numpy==1.26.4
tldextract==5.1.1

# --- MACHINE LEARNING & DEEP LEARNING ---
xgboost==2.0.3
tensorflow==2.15.0
scikit-learn==1.4.1.post1

# --- ADVANCED NETWORK FEATURES (WHOIS, DNS, SIMILARITY) ---
python-whois==0.8.0
dnspython==2.6.1
python-Levenshtein==0.25.0
requests==2.31.0

# --- DATABASE & ENVIRONMENT ---
supabase==2.4.0
python-dotenv==1.0.1

# --- UTILS ---
pydantic==2.6.3